String Formatting with %, Part 1
 Blessed [is] he that readeth, and they that hear the words of this prophecy, and keep those things which are written therein: for the time [is] at hand. (Revelation 1:3)

STRINGS & CONSOLE OUTPUT
String Formatting with %, Part 1
When you want to print a variable with a string, there is a better method than concatenating strings together.

name = "Mike"
print "Hello %s" % (name)
The % operator after a string is used to combine a string with variables. The % operator will replace a %s in the string with the string variable that comes after it.
Instructions
1.
Take a look at the code in the editor. What do you think it'll do? Click Run when you think you know.