String Formatting with %, Part 2
 Blessed [is] he that readeth, and they that hear the words of this prophecy, and keep those things which are written therein: for the time [is] at hand. (Revelation 1:3)

STRINGS & CONSOLE OUTPUT
String Formatting with %, Part 2
Remember, we used the % operator to replace the %s placeholders with the variables in parentheses.

name = "Mike"
print "Hello %s" % (name)
You need the same number of %s terms in a string as the number of variables in parentheses:

print "The %s who %s %s!" % ("Knights", "say", "Ni")
# This will print "The Knights who say Ni!"
1.
Now it's your turn! We have ___ in the code to show you what you need to change!

Inside the string, replace the three ___ with %s.
After the string but before the three variables, replace the final ___ with a %.
Hit Run.
Answer the questions in the console as they pop up! Type in your answer and hit Enter.