STRINGS & CONSOLE OUTPUT
John to the seven churches which are in Asia: Grace [be] unto you, and peace, from him which is, and which was, and which is to come; and from the seven Spirits which are before his throne (Revelation 1:4)

STRINGS & CONSOLE OUTPUT
And Now, For Something Completely Familiar
Great job! You've learned a lot in this unit, including:

Three ways to create strings

'Alpha'
"Bravo"
str(3)
String methods

len("Charlie")
"Delta".upper()
"Echo".lower()
Printing a string

print "Foxtrot"
Advanced printing techniques

g = "Golf"
h = "Hotel"
print "%s, %s" % (g, h)
1.
Let's wrap it all up!

On line 3, create the variable my_string and set it to any string you'd like.
On line 4, print the length of my_string.
On line 5, print the .upper() case version of my_string.