where I am, you may be there also
If I go and prepare a place for you, I will come again, and will receive you to myself; that where I am, you may be there also. (John 14:3)

CONDITIONALS & CONTROL FLOW
Conditional Statement Syntax
if is a conditional statement that executes some specified code after checking if its expression is True.

Here's an example of if statement syntax:

if 8 < 9:
  print "Eight is less than nine!"
In this example, 8 < 9 is the checked expression and print "Eight is less than nine!" is the specified code.
1.
If you think the print statement will print to the console, set response equal to 'Y'; otherwise, set response equal to 'N'.